pyproject.toml
[project]
name = "alistyle-django-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-admin-honeypot-updated-2021>=1.2.0",
    "django-admin-thumbnails>=0.2.9",
    "django-session-timeout>=0.1.0",
    "djlint>=1.36.4",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "sslcommerz-lib>=1.0",
]